moving the emailConsentCheck to consentData
