+++ .idea/workspace.xml
@@ -5,15 +5,10 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="f5614ad5-0748-41e2-a0c7-81b3af1be12f" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/pages/api/emailSubscription.js" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/pages/unsubscribe.jsx" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/pages/account.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/pages/account.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pages/api/emailConsentCheck.js" beforeDir="false" afterPath="$PROJECT_DIR$/pages/api/emailConsentCheck.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pages/api/emailData.js" beforeDir="false" afterPath="$PROJECT_DIR$/pages/api/emailData.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pages/api/emailSent.js" beforeDir="false" afterPath="$PROJECT_DIR$/pages/api/emailSent.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pages/youPhished.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/pages/youPhished.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/templates/facebook.handlebars" beforeDir="false" afterPath="$PROJECT_DIR$/templates/facebook.handlebars" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/pages/api/consentData.js" beforeDir="false" afterPath="$PROJECT_DIR$/pages/api/consentData.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/pages/api/emailConsentCheck.js" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
